Set slimmer headers to remove meta viewport
For every controller, we should be setting the meta viewport removal
header for slimmer so that the post-processing of the views occur.

Also renamed the method currently being used 'set_analytics_headers', to
'set_app_slimmer_headers' so that all of the slimmer header setters can
be contained in one place.