Highlight consigned commodities in overview.
If commodity description is phrased like 'Consigned from <country>' adds
special block to highlight that it is consigned.

NOTE: as far as I can tell there is no other, more intelligent way to
identify such commodities. There is no identification in the database and
word 'consigned' is not present in Taric's manual.