Remove pjax
We added pjax to help improve the perceived application load time, but
with the caching that is in place this really isn't needed and just
adds more complexity with JS and debugging JS issues.

Change GOVUK.tariff to have one onLoad function and move searchForm.initialize()
into that function.

removes all other references to pjax links