Remove router rake task
This is a hangover of the old router. Trade Tariff doesn't register
against the new router (which is a good thing because it wouldn't work
as it would register in preview with a domain of
production.alphagov.co.uk because it's based on FACTER_govuk_platform)

I'm removing FACTER_govuk_platform and I stumbled across this, so I
thought I'd just delete the rake task.