Ensure old test runs get cleaned up
CI builds were failing prior to this because we'd run
`assets:precompile` on an earlier run and it left old files around
before running the tests.

The JS has totally changed as of f6ed14d, so we need to ensure we get
the fresh JS before running specs.

Verified this works by blowing away the workspace and re-running the
tests on CI - they now pass.