Switch off classic analytics
We're moving over to universal analytics a few apps at a time. This is
the first app to go over. We're keeping the classic tracking code but
sending it to universal analytics instead.
This work should only be merged after this
https://github.com/alphagov/govuk_frontend_toolkit/pull/190 has been
merged, and the front end toolkit has been updated.